requirements.txt
langchain
transformers
pdfminer.six